Remove redundant `--disable-aligned-segments` options
This option became the default in versions:
GASNet 1.30 (b72d89a6b28) / 2017.12.0 (db6472e7d4),
and has always been required for use of PSHM (--enable-pshm).

The aligned-segments feature is deprecated and only preserved for legacy reasons,
and is likely to eventually be removed entirely (i.e. there is no
chance it will ever be default-enabled again).