Remove obsolescent `---with-segment-mmap-max` options
The `--with-segment-mmap-max` option was removed in:
GASNet 1.32.0 (853e522dca) / 2018.12.0 (57c3834ddb)
Since then, the option does nothing but generate a configure warning
about use of the obsolete option.

Luckily the mechanism for determining the default mmap max received a
major overhaul in the versions listed above, and is now smart enough to
work "out of the box" for nearly all systems. So although there is a new
configure option (with different spelling) to manually override that
defaulting algorithm, it is best NOT to pass that new option unless you
encounter a system-specific problem with the segment probe.