Configure GPU memory kinds when Legion has GPUs enabled.
This commit follows the variables used in Legion's Makefile system
under `runtime/runtime.mk`. These two variables control the build:

  * `USE_CUDA=1` causes Legion to be built with CUDA support
  * `USE_HIP=1` causes Legion to be built with HIP support

We expect that the vast majority of users who build Legion will want
the corresponding GASNet GPU memory kind to be enabled if Legion
itself has GPUs enabled. Therefore, we turn this on if we find that
these Legion build variables have been set.

To disable or override this feature build with the environment
variable GASNET_GPU_CONFIGURE_ARGS set to an empty string (or any
other value of the user's choice).